Add files via upload
This data analysis was perform for my course assignment with UC Davis via Coursera. I was using a dataset from a US-based organization called Yelp, which provides a platform for users to provide reviews and rate their interactions with a variety of organizations – businesses, restaurants, health clubs, hospitals, local governmental offices, charitable organizations, etc. Yelp has made a portion of this data available for personal, educational, and academic purposes. Also, be sure to check out the Yelp Dataset ER Diagram and instructions below for more detailed information on the dataset.